chore: add MIT License file
Add a new file named LICENSE with the content of the MIT License. This license grants permission to any person obtaining a copy of the software to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, subject to the conditions specified in the license. The license also includes a disclaimer of warranty and limits the liability of the authors or copyright holders.